Implement localStorage and fixes bug with inbox
The user's projects and inbox will be stored on page unload, and they
will be loaded and displayed on page load.

Additionally, all data in inbox were stored in an object outside of
projectArray, which meant that it was not loading or displaying on page
load. The data was shifted to the right place and the bug was fixed.